add face alignment

+++ include/face_align.h
@@ -8,11 +8,11 @@
 
 class FaceAlign { 
   private:
-    std::vector<cv::Point2d> align_src_;
+    std::vector<cv::Point2f> align_src_;
     cv::Size size;
   public:
     FaceAlign();
-    cv::Mat Align(cv::Mat& input, const std::vector<cv::Point2d>& align_dst);
+    cv::Mat Align(cv::Mat& input, const std::vector<cv::Point2f>& align_dst);
 };
 
 #endif
